Inserted the new table for balances. Added the logic to calculate balances. Changed the data type from float to double for amounts

+++ app/src/main/java/com/kinetx/moneymanager/database/TransactionDatabase.kt
@@ -12,7 +12,7 @@ data class TransactionDatabase(
     val transactionId : Long = 0L,
 
     @ColumnInfo(name = "amount")
-    val transactionAmount : Float = 0.0f,
+    val transactionAmount : Double = 0.0,
 
     @ColumnInfo(name="transaction_type")
     val transactionType : TransactionType,
